fix: address CodeRabbit review findings on PR #245
- Move swh_ajax_send_test_email() from bootstrap to admin/class-settings.php
  (admin-only handler belongs in admin module, not the thin bootstrap)
- Send test email to requesting admin's own email (wp_get_current_user()),
  falling back to admin_email option — fixes multi-admin UX (#103)
- Add [3.1.0] compare link and retarget [Unreleased] in CHANGELOG.md
- Add = 3.1.0 = upgrade notice entry to readme.txt
- Move inline `import re` to module level in test_helpdesk_pw.py
- Remove duplicate _swh_unread assertion in test_50_unread_badge
